+++ src/ecotally/__init__.py
@@ -1,9 +1,9 @@
 """Small, transparent tools for community ecology."""
 
-__version__ = "0.15.0"
+__version__ = "0.16.0"
 
 from .beta import PairwiseResult, compare_communities
-from .diversity import DiversityResult, calculate_diversity
+from .diversity import DiversityResult, calculate_diversity, hill_number
 from .estimation import RichnessEstimate, estimate_richness, rarefaction_curve
 from .summary import DatasetSummary, summarize_dataset, summarize_species
 from .quality import QualityIssue, audit_communities
@@ -18,6 +18,7 @@ __all__ = [
     "DiversityResult",
     "PairwiseResult",
     "calculate_diversity",
+    "hill_number",
     "compare_communities",
     "RichnessEstimate",
     "estimate_richness",